requirements.txt
distro==1.8.0 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:02e111d1dc6a50abb8eed6bf31c3e48ed8b0830d1ea2a1b78c61765c2513fdd8 \
    --hash=sha256:99522ca3e365cac527b44bde033f64c6945d90eb9f769703caaec52b09bbd3ff
gitdb==4.0.10 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:6eb990b69df4e15bad899ea868dc46572c3f75339735663b81de79b06f17eb9a \
    --hash=sha256:c286cf298426064079ed96a9e4a9d39e7f3e9bf15ba60701e95f5492f28415c7
gitpython==3.1.32 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:8d9b8cb1e80b9735e8717c9362079d3ce4c6e5ddeebedd0361b228c3a67a62f6 \
    --hash=sha256:e3d59b1c2c6ebb9dfa7a184daf3b6dd4914237e7488a1730a6d8f6f5d0b4187f
httplib2==0.22.0 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:14ae0a53c1ba8f3d37e9e27cf37eabb0fb9980f435ba405d546948b009dd64dc \
    --hash=sha256:d7a10bc5ef5ab08322488bde8c726eeee5c8618723fdb399597ec58f3d82df81
launchpadlib==1.11.0 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:01898c937477b0c64a75338adb0977028d7346a8a019eb023cf68fed99850146 \
    --hash=sha256:d71b4ff87184e0c54735cec24efb5a2bbd114c771caf131b9f1b6b343cf4eb29
lazr-restfulclient==0.14.5 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:0751717c7e74db1987e9a77335707d4d7d97cf04b1ad0898b822f12333d6887c \
    --hash=sha256:5ac92f7b2238cbcb54a9710830826d6140b74a0f9f4100cc713c1d872a6e766e \
    --hash=sha256:6ed6010b354cd15aa0fdcdc89365d07d933565dbb6907abffff17729399875fb
lazr-uri==1.0.6 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:5026853fcbf6f91d5a6b11ea7860a641fe27b36d4172c731f4aa16b900cf8464
oauthlib==3.2.2 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:8139f29aac13e25d502680e9e19963e83f16838d48a0d71c287fe40e7067fbca \
    --hash=sha256:9859c40929662bec5d64f34d01c99e093149682a3f38915dc0655d5a633dd918
pyparsing==3.1.1 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:32c7c0b711493c72ff18a981d24f28aaf9c1fb7ed5e9667c9e84e3db623bdbfb \
    --hash=sha256:ede28a1a32462f5a9705e07aea48001a08f7cf81a021585011deba701581a0db
setuptools==68.1.2 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:3d4dfa6d95f1b101d695a6160a7626e15583af71a5f52176efa5d39a054d475d \
    --hash=sha256:3d8083eed2d13afc9426f227b24fd1659489ec107c0e86cec2ffdde5c92e790b
six==1.16.0 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:1e61c37477a1626458e36f7b1d82aa5c9b094fa4802892072e49de9c60c4c926 \
    --hash=sha256:8abb2f1d86890a2dfb989f9a77cfcfd3e47c2a354b01111771326f8aa26e0254
smmap==5.0.0 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:2aba19d6a040e78d8b09de5c57e96207b09ed71d8e55ce0959eeee6c8e190d94 \
    --hash=sha256:c840e62059cd3be204b0c9c9f74be2c09d5648eddd4580d9314c3ecde0b30936
wadllib==1.3.6 ; python_full_version > "3.8.0" and python_version < "4.0" \
    --hash=sha256:acd9ad6a2c1007d34ca208e1da6341bbca1804c0e6850f954db04bdd7666c5fc
